// src/api/subscription/api.ts
export interface BillingOptionDto {
  /**
   *
   * @type {string}
   * @memberof BillingOptionDto
   */
  interval?: BillingOptionDtoIntervalEnum;
  /**
   *
   * @type {number}
   * @memberof BillingOptionDto
   */
  price?: number;
  /**
   *
   * @type {string}
   * @memberof BillingOptionDto
   */
  currency?: string;
  /**
   *
   * @type {string}
   * @memberof BillingOptionDto
   */
  priceId?: string;
}

export const BillingOptionDtoIntervalEnum = {
  Monthly: 'monthly',
  Annually: 'annually',
} as const;

export type BillingOptionDtoIntervalEnum =
  (typeof BillingOptionDtoIntervalEnum)[keyof typeof BillingOptionDtoIntervalEnum];

/**
 *
 * @export
 * @interface CancelSubscriptionResponse
 */
export interface CancelSubscriptionResponse {
  /**
   *
   * @type {string}
   * @memberof CancelSubscriptionResponse
   */
  subscriptionId?: string;
  /**
   *
   * @type {string}
   * @memberof CancelSubscriptionResponse
   */
  canceledAt?: string;
  /**
   *
   * @type {string}
   * @memberof CancelSubscriptionResponse
   */
  status?: CancelSubscriptionResponseStatusEnum;
}

export const CancelSubscriptionResponseStatusEnum = {
  Canceled: 'canceled',
  Error: 'error',
} as const;

export type CancelSubscriptionResponseStatusEnum =
  (typeof CancelSubscriptionResponseStatusEnum)[keyof typeof CancelSubscriptionResponseStatusEnum];

/**
 *
 * @export
 * @interface CheckPaymentStatusRequest
 */
export interface CheckPaymentStatusRequest {
  /**
   *
   * @type {string}
   * @memberof CheckPaymentStatusRequest
   */
  sessionId: string;
}
/**
 *
 * @export
 * @interface CheckPaymentStatusResponse
 */
export interface CheckPaymentStatusResponse {
  /**
   *
   * @type {string}
   * @memberof CheckPaymentStatusResponse
   */
  sessionId?: string;
  /**
   *
   * @type {string}
   * @memberof CheckPaymentStatusResponse
   */
  status?: CheckPaymentStatusResponseStatusEnum;
  /**
   *
   * @type {string}
   * @memberof CheckPaymentStatusResponse
   */
  subscriptionId?: string;
  /**
   *
   * @type {string}
   * @memberof CheckPaymentStatusResponse
   */
  customerEmail?: string;
}

export const CheckPaymentStatusResponseStatusEnum = {
  Complete: 'complete',
  Open: 'open',
  Expired: 'expired',
} as const;

export type CheckPaymentStatusResponseStatusEnum =
  (typeof CheckPaymentStatusResponseStatusEnum)[keyof typeof CheckPaymentStatusResponseStatusEnum];

/**
 *
 * @export
 * @interface CreateCheckoutRequest
 */
export interface CreateCheckoutRequest {
  /**
   *
   * @type {string}
   * @memberof CreateCheckoutRequest
   */
  priceId: string;
}
/**
 *
 * @export
 * @interface CreateCheckoutSessionResponse
 */
export interface CreateCheckoutSessionResponse {
  /**
   *
   * @type {string}
   * @memberof CreateCheckoutSessionResponse
   */
  clientSecret?: string;
}
/**
 *
 * @export
 * @interface ErrorResponse
 */
export interface ErrorResponse {
  /**
   *
   * @type {string}
   * @memberof ErrorResponse
   */
  error?: string;
  /**
   *
   * @type {number}
   * @memberof ErrorResponse
   */
  code?: number;
}
/**
 *
 * @export
 * @interface GetAllSubscriptionHistoryPaginatedResponse
 */
export interface GetAllSubscriptionHistoryPaginatedResponse {
  /**
   *
   * @type {boolean}
   * @memberof GetAllSubscriptionHistoryPaginatedResponse
   */
  hasMore?: boolean;
  /**
   *
   * @type {Array<SubscriptionHistoryResponse>}
   * @memberof GetAllSubscriptionHistoryPaginatedResponse
   */
  data?: Array<SubscriptionHistoryResponse>;
}
// ...
export interface SubscriptionHistoryResponse {
  /**
   *
   * @type {string}
   * @memberof SubscriptionHistoryResponse
   */
  subscriptionId?: string;
  /**
   *
   * @type {string}
   * @memberof SubscriptionHistoryResponse
   */
  invoiceId?: string;
  /**
   *
   * @type {string}
   * @memberof SubscriptionHistoryResponse
   */
  planName?: string;
  /**
   *
   * @type {number}
   * @memberof SubscriptionHistoryResponse
   */
  price?: number;
  /**
   *
   * @type {string}
   * @memberof SubscriptionHistoryResponse
   */
  currency?: string;
  /**
   *
   * @type {string}
   * @memberof SubscriptionHistoryResponse
   */
  billingDate?: string;
  /**
   *
   * @type {string}
   * @memberof SubscriptionHistoryResponse
   */
  status?: SubscriptionHistoryResponseStatusEnum;
  /**
   *
   * @type {string}
   * @memberof SubscriptionHistoryResponse
   */
  downloadUrl?: string;
}

export const SubscriptionHistoryResponseStatusEnum = {
  Active: 'ACTIVE',
  Canceled: 'CANCELED',
  Incomplete: 'INCOMPLETE',
  IncompleteExpired: 'INCOMPLETE_EXPIRED',
  PastDue: 'PAST_DUE',
  Trialing: 'TRIALING',
  Unpaid: 'UNPAID',
} as const;

export type SubscriptionHistoryResponseStatusEnum =
  (typeof SubscriptionHistoryResponseStatusEnum)[keyof typeof SubscriptionHistoryResponseStatusEnum];

/**
 *
 * @export
 * @interface SubscriptionPlanDto
 */
export interface SubscriptionPlanDto {
  /**
   *
   * @type {string}
   * @memberof SubscriptionPlanDto
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof SubscriptionPlanDto
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof SubscriptionPlanDto
   */
  price?: string;
  /**
   *
   * @type {string}
   * @memberof SubscriptionPlanDto
   */
  currency?: string;
  /**
   *
   * @type {string}
   * @memberof SubscriptionPlanDto
   */
  interval?: string;
  /**
   *
   * @type {string}
   * @memberof SubscriptionPlanDto
   */
  description?: string;
  /**
   *
   * @type {boolean}
   * @memberof SubscriptionPlanDto
   */
  recommended?: boolean;
  /**
   *
   * @type {Array<string>}
   * @memberof SubscriptionPlanDto
   */
  features?: Array<string>;
  /**
   *
   * @type {Array<BillingOptionDto>}
   * @memberof SubscriptionPlanDto
   */
  billingOptions?: Array<BillingOptionDto>;
}
/**
 *
 * @export
 * @interface SubscriptionPlansResponse
 */
export interface SubscriptionPlansResponse {
  /**
   *
   * @type {Array<SubscriptionPlanDto>}
   * @memberof SubscriptionPlansResponse
   */
  data?: Array<SubscriptionPlanDto>;
}
/**
 *
 * @export
 * @interface UserSubscriptionPlanResponse
 */
export interface UserSubscriptionPlanResponse {
  /**
   *
   * @type {string}
   * @memberof UserSubscriptionPlanResponse
   */
  planName?: string;
  /**
   *
   * @type {string}
   * @memberof UserSubscriptionPlanResponse
   */
  billingPeriod?: UserSubscriptionPlanResponseBillingPeriodEnum;
  /**
   *
   * @type {number}
   * @memberof UserSubscriptionPlanResponse
   */
  price?: number;
}

export const UserSubscriptionPlanResponseBillingPeriodEnum = {
  Monthly: 'monthly',
  Annually: 'annually',
} as const;

export type UserSubscriptionPlanResponseBillingPeriodEnum =
  (typeof UserSubscriptionPlanResponseBillingPeriodEnum)[keyof typeof UserSubscriptionPlanResponseBillingPeriodEnum];
// ...
export class SubscriptionApi extends BaseAPI {
  /**
   *
   * @summary Cancel a subscription for a client
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SubscriptionApi
   */
  public cancelSubscription(options?: RawAxiosRequestConfig) {
    return SubscriptionApiFp(this.configuration)
      .cancelSubscription(options)
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Check Stripe Checkout Session status
   * @param {CheckPaymentStatusRequest} checkPaymentStatusRequest
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SubscriptionApi
   */
  public checkPaymentStatus(checkPaymentStatusRequest: CheckPaymentStatusRequest, options?: RawAxiosRequestConfig) {
    return SubscriptionApiFp(this.configuration)
      .checkPaymentStatus(checkPaymentStatusRequest, options)
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Create a Stripe Checkout Session
   * @param {CreateCheckoutRequest} createCheckoutRequest
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SubscriptionApi
   */
  public createCheckoutSession(createCheckoutRequest: CreateCheckoutRequest, options?: RawAxiosRequestConfig) {
    return SubscriptionApiFp(this.configuration)
      .createCheckoutSession(createCheckoutRequest, options)
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Get current subscription plan for a client
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SubscriptionApi
   */
  public getCurrentPlan(options?: RawAxiosRequestConfig) {
    return SubscriptionApiFp(this.configuration)
      .getCurrentPlan(options)
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Get paginated subscription history
   * @param {string} [from] The starting index of subsciption, you should pass last subscription id from the previous page.
   * @param {number} [size] The number of invoices to return.
   * @param {GetSubscriptionHistoryStatusEnum} [status] Filter by subscription status.
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SubscriptionApi
   */
  public getSubscriptionHistory(
    from?: string,
    size?: number,
    status?: GetSubscriptionHistoryStatusEnum,
    options?: RawAxiosRequestConfig
  ) {
    return SubscriptionApiFp(this.configuration)
      .getSubscriptionHistory(from, size, status, options)
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Get available subscription plans
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof SubscriptionApi
   */
  public getSubscriptionPlans(options?: RawAxiosRequestConfig) {
    return SubscriptionApiFp(this.configuration)
      .getSubscriptionPlans(options)
      .then((request) => request(this.axios, this.basePath));
  }
}

/**
 * @export
 */
export const GetSubscriptionHistoryStatusEnum = {
  Active: 'active',
  Canceled: 'canceled',
  Incomplete: 'incomplete',
  IncompleteExpired: 'incomplete_expired',
  PastDue: 'past_due',
  Trialing: 'trialing',
  Unpaid: 'unpaid',
} as const;
export type GetSubscriptionHistoryStatusEnum =
  (typeof GetSubscriptionHistoryStatusEnum)[keyof typeof GetSubscriptionHistoryStatusEnum];